A fix (bug #7129: Test failure: 'type_ranges' when using '--ps-protocol'        bug #7126: Test failure: 'func_str' when using '--ps-protocol'        bug #7130: Test failure: 'type_uint' when using '--ps-protocol').

+++ libmysql/libmysql.c
@@ -3663,8 +3663,8 @@ static void fetch_result_with_conversion(MYSQL_BIND *param, MYSQL_FIELD *field,
   case MYSQL_TYPE_INT24: /* mediumint is sent as 4 bytes int */
   case MYSQL_TYPE_LONG:
   {
-    long value= sint4korr(*row);
-    longlong data= field_is_unsigned ? (longlong) (unsigned long) value :
+    int32 value= sint4korr(*row);
+    longlong data= field_is_unsigned ? (longlong) (uint32) value :
                                        (longlong) value;
     fetch_long_with_conversion(param, field, data);
     *row+= 4;
